+++ src/detector.js
@@ -27,7 +27,6 @@ const MODEL_FILES = [
   'ssd_mobilenetv1_model-shard2',
   'face_landmark_68_model-weights_manifest.json',
   'face_landmark_68_model-shard1',
-  'face_landmark_68_model-shard2',
   'face_recognition_model-weights_manifest.json',
   'face_recognition_model-shard1',
   'face_recognition_model-shard2',